refactor: NsUser가 강의에 등록한 NsStudent 생성
