+++ src/main/java/nextstep/courses/CannotCreateSessionException.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses;
 
 public class CannotCreateSessionException extends RuntimeException {
     public CannotCreateSessionException(String message) {
