+++ src/main/java/nextstep/courses/CannotRegisterException.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses;
 
 public class CannotRegisterException extends RuntimeException {
     public CannotRegisterException(String message) {
