+++ src/main/java/nextstep/courses/CannotUploadImageException.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses;
 
 public class CannotUploadImageException extends RuntimeException {
     public CannotUploadImageException(String message) {
