+++ src/main/java/nextstep/courses/InvalidNaturalNumberException.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses;
 
 public class InvalidNaturalNumberException extends RuntimeException {
     public InvalidNaturalNumberException(String message) {
