+++ src/main/java/nextstep/courses/InvalidPeriodException.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses;
 
 public class InvalidPeriodException extends RuntimeException {
     public InvalidPeriodException(String message) {
