+++ src/test/java/nextstep/courses/domain/ImageTypeTest.java
@@ -1,5 +1,6 @@
 package nextstep.courses.domain;
 
+import nextstep.courses.CannotUploadImageException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
